Move `parse_cfg_attr` to rustc_attr_parsing
